requirements.txt
# Production-ready requirements.txt - Core dependencies only

# Core FastAPI and web framework
fastapi==0.115.13
uvicorn==0.34.3
starlette==0.46.2
python-multipart==0.0.20

# Database and ORM
sqlalchemy==2.0.41
alembic==1.16.2
psycopg2-binary==2.9.10

# Authentication and Security
passlib==1.7.4
python-jose==3.5.0
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==45.0.4

# Data processing and file handling
pandas==2.3.0
numpy==2.3.1
openpyxl==3.1.5

# Rate limiting
slowapi==0.1.9

# System monitoring
psutil

# Validation and schemas  
pydantic==2.11.7
pydantic_core==2.33.2
email_validator==2.2.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.9.0.post0

# Security enhancements
bleach
defusedxml

# Core dependencies (auto-installed but listed for clarity)
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
et_xmlfile==2.0.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
limits==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pyasn1==0.6.1
pycparser==2.22
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2